requirements.txt
constructs>=10.0.0,<11.0.0
attrs==23.1.0
aws-cdk-lib==2.115.0
aws-cdk.asset-awscli-v1==2.2.201
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.1
aws-cdk.aws-codestar-alpha==2.115.0a0
cattrs==23.2.3
importlib-resources==6.1.1
jsii==1.93.0
publication==0.0.3
python-dateutil==2.8.2
six==1.16.0
typeguard==2.13.3
typing_extensions==4.9.0
